Fixed bug introduced to IsPonyOnScreen now actually using the location parameter.
Fixed set operation of Behavior property to only set allowable directions when not in a game.
Fixed bug in DesktopPonyAnimator where removing all sprites may not return to the menu (race condition, now check on every loop to be sure).
Removed unessessary ByVal keywords.
Removed ByRef keywords where they were not needed and the parameter reference was not changed.