Removed SplitContainer from PonyEditor as it required Begin/EndInit block to work, and this interface is not implemented in the Mono WinForms port.
Changed ExceptionDialog to use a fixed size for the textbox on Mono, as the GetpreferredSize method returns a uselessly small size otherwise.
Added Application.DoEvents to GameSelectionForm so Mono would draw it all before loading icons.
Changed IdleWorker to fallback to BeginInvoke queue under Mono, as the Idle event is not raised reliably.
Fixed pony logs only being enabled in debug builds.
Fixed destinationCoords to not be set when a game is active, as it overrides them.