Fixed a threading issue that was exposed by calling Draw from within PonyEditorAnimation on a different thread. This function has been made private as it need not be exposed, and this will prevent potentially silly calls.
Made the setting of the sprites field use Interlocked.Exchange to ensure the rendering thread would see the update.
Corrected DisplayBounds to not call render, as the sprites list would possibly be changed and thus it would be wrong to redraw from it. Changed doc to note that setting the property would clear the form.