Added CaseInsensitiveString struct and applied to IPonyIniSerializable.Name. Using this should help ensure the correct type of string comparison is done in more places. Had to refactor a bunch of code to allow for this change. Found some code doing a case-insenstive comparison where a case-sensitive comparison should have been used, and so fixed those.
Effects can no longer be specified as follow targets in .ini files due to naming ambiguity issues. Internally, an effect may be manually set as a target (for use in games).
Added a Count property to PonySelectionControl so the constant reparsing of the pony count text can be avoided. Added some additional checks to limit range of values that can be used.
Changed editor to re-determine the link order when linked behaviors are changed. Sped up this code as it was on a hot path - main improvement providing the string to the RowCollection.Add method rather than setting after the fact as the grid's internal faffing with cached rows makes setting really slow.
Added OnlyOrDefault LINQ methods - like SingleOrDefault except providing the default value in the both non-single cases, rather than throwing if more than one match is found.