Introduced Pony2 class - a rebuild of the Pony class from scratch. This is intended to try and remove the extremely high and confusing coupling with other areas by starting with a private state, and gradually reintroducing more structured access to internals as needed. This has informed several interfaces used in the previous dozen or so commits.
To facilitate it, a PonyContext class has been introduced, intended to provide a more focused collection of external state that individual pony sprites will rely upon. The Effect class has been modified somewhat to allow both the Pony and Pony2 class to work with it. Similarly the interaction class has been made generic to assist with this.