Fixed a regression concerning hit testing for draggable sprites.
Hit testing has always been done by considering the nearest element to a click within the bounding box of the sprite, ignoring the actual sprite pixels. This means the wrong sprite can be selected for dragging if two are very close but it's an acceptable trade since it makes hit testing much easier, and usually the closeness means the user won't notice anyway.
Before the common IDraggableSprite interface came about, ponies and effects were tested separately for hit testing, all ponies first and then all effects. An accidental side effect of this was to apply an ordering to hit testing - preferring ponies within bounds of the click over effects. Combined with the approximate hit testing, attempting to click a pony 'through' a transparent effect would usually start dragging the pony. This is actually useful if entirely accidental behavior.
When the common interface removed this accidental ordering of hit testing, a regression occurred where transparent effects would often get dragged since now the test was merely done for closest draggable sprite.
Although approximating the hit testing between two ponies or two effects and being wrong sometimes isn't much of an issue, the unintended behavior of preferring ponies over effects when hit testing between them was useful, and losing it was a regression.
To fix it, ponies are now tested first, and then any draggable sprite is considered.