Polished the new editor interface. It's also no longer restricted to debug builds.
Added some basic icons to add some color, and mostly to help provide hints when buttons have become activated (the de-colorization of the image makes it obvious the button is disabled, whereas with text only it is very hard to tell).
Making changes will cause the main menu to reload on return.
Tweaked new items to not be marked dirty immediately.
Fixed alignment of grid view on interactions viewer.
The ponies node will now be marked OK once validation completes.
Hitting the enter key is the documents tree will now open item collection tabs as well as item tabs.
Right clicking on item collections or items in the tree view now presents a right click menu with a single "Open" command to open the tab.
Added access to the image operation forms from the behavior editor.