The Mac platform does not handle the multiple UI thread setup the current architecture uses to have all the WinForms menu running on the program main thread, and then use a second UI thread for rendering. This is because Mac has a concept of a run loop. The main thread of the program has one installed for free, but other threads require this loop to be started manually. The platform invoking required is too complex to try that, and attempting to piggyback the Gtk loop within the WinForms loop is also not possible.
Instead, the program will be redesigned to run Gtk on the program main loop, and share that thread for both menus and ponies alike. This means the existing WinForms menus cannot be used sadly, but a simple Gtk selection menu has been made and this is sufficient for selecting and running ponies. Options and editing ponies are not vital.

The entry point to the program has been changed to spin up the new Gtk menu on Mac, and otherwise spin up the usual WinForms menu on Windows/Unix.

A bunch of Mac specific workarounds and hacks in the WinForms layer has been removed since it can no longer be triggered. The right click menu in DesktopPonyAnimator has been refined to remove enteries that are restricted by platform.