Moved downscaling code into the Desktop Ponies assembly, and added property on interfaces for specifying a buffer preprocessing function. This means the downscaling can now be applied for the Gtk interface, and well as on the main menu, netting some nice memory savings.
The GifDecoder no longer provides a hash when decoding buffers, instead a generic hash function is provided on the non-generic GifImage which can be called where needed. This also means hashing can be delayed until after downscaling has been applied, which should reduce the hash time a bit as a result.