Fixed some rounding issues with image centering. The form now correctly mirrors by subtracting one from the image width. It also draws the marker in a way such that the pixel can actually be seen.
A RoundingPolicy has been added to SpriteImage to control how the center pixel is determined for images with even width or height. In this case the center straddles two pixels. Separate policies apply for X and Y directions, this allows the left and right images to have a policy that mirrors the X coordinate (Floor and Ceiling) but keep the Y coordinate in sync (with the default rounding mode of ToEven).