Ensure the current directory of the application matches the directory containing the executable (for example, when started by an external tool like the CLRProfiler, the current directory will instead be inherited from that in use by the tool) which allows the EvilGlobals.InstallLocation property to be removed as we can assume the current directory will be correct.
The means when starting as a screensaver we can do without the hacky workaround for correcting the path and thus we simplify the whole process since we can now rely on relative paths throughout the application being relative to the executable and thus looking up the location of other resources correctly.

This also saves some memory since shorter relative paths will be stored rather than absolute paths. Another memory saving is achieved by marking the FollowOffsetType enum as a byte.