Significant code cleanup to the editor and related dialogs. Refactored a lot of duplicated code into reusable methods. As a result, made validation more consistent between the new item dialogs and when editing via the grid views.
Changed effect duplicate name validation to only consider effects of the behavior and not all effects. That was a holdover from when following effects was possible and thus all names needed to be unique. This is no longer possible so this restriction is not needed.

Fixed row deletion handlers to reload the pony after deleting a row so this functionality works again.

Fixed extracted SetImage method (now in EditorCommon) to dispose of the image that was loaded to check it was valid. Previously it was not disposed and thus the image object would keep a handle to the file open and cause other spurious errors.

Tidied up the design of the new item dialogs to make a bit more use of the available space and also to allow resizing. Also provided better input for some values and defaulted some dropdowns to reduce the amount of work required to create a a new item.

Warn when attempting to overwrite an existing file when copying in new files. This is to prevent the user accidentally overwriting the files. Additionally, the preview will hold open the file handle which means we may not be able to alter the file anyway because our process is locking it.