Fixed ponies being able to escape the zone and trigger teleportation if the had a high speed and small image and thus made it fully outside in one step. If they were previously in zone, the rebounding will be applied rather than letting the teleportation failsafe kick in. If they are still outside come next step the failsafe will then kick in.
Similarly, if a sudden change in image size occurs then leaves the pony partially out of zone, they will now be permitted to run that behavior to completion before the return to zone destination will be set forcing them back into the area. If they can move this will hopefully be sufficient time for them to return to the zone. If not then their stationary animation can play out before a moving animation resumes. This is handy for transition animations like warping in after transit so that the flow of animations is not broken.