Removed events using CollectionItemChangedEventArgs from AnimationLoopBase, preferring to instead call on the collection based event and provide the single element in a wrapped collection. This simplifies life for consumers since they cannot accidentally miss subscription to either event.
This fixes a bug in Editor2PonyAnimator which had subscribed to the single event, and would no longer update the info about the preview pony. This was because the pony was being add via the bulk insert and thus the collection based event was being raised with a single element instead. Now there is only a single pair of events to worry about this cannot happen in future.