Fix a bug in UnscalableImagesForm where the background loading would attempt to call SmartInvoke after the form had been closed.
To mitigate this problem in future, SmartInvoke will now perform a disposed check on the control. This prevents calls from the UI thread executing for a disposed control. Unfortunately for cross thread calls a race condition still exists. If the control is disposed after the call to InvokeRequired, an exception is thrown from BeginInvoke as a window handle cannot be found. Worse, if the control is disposed before the call to BeginInvoke then InvokeRequired will return false and will not throw - meaning we will try and directly execute the method on a thread other than the UI thread for a control that is disposed. This is an incredibly unfortunate race condition that cannot easily be mitigated.