Changed interactions to allow each target to individually select any behavior from those listed, subject to group restrictions, and run that.
This affects the 'Any' mode of interaction by relaxing the restriction that all targets considered must be able to start a common behavior. If one did not exist previously the interaction did not run, now instead targets unable to run a behavior simply won't participate.

For interactions with multiple listed behaviors, this allows targets to choose a behavior individually instead of having to start a common behavior. This means interactions can be engineered to allow targets to start different behaviors depending what pony they are, or their current group.