Play speech when entering the sleep or drag states.
If a behavior has been specifically marked as being used for sleep or dragging, allow the speech attached to it to be played. This way if a behavior has been specifically made for these scenarios any speech for it can be used. If a fallback behavior is used because a specific behavior for these states does not exist, we won't play a speech since it may not be appropriate.